+++ platform/src/codatplatform/models/operations/list_api_keys.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import apikeys as shared_apikeys
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import apikeys as shared_apikeys
 from typing import Optional
 
 
@@ -12,13 +11,11 @@ from typing import Optional
 class ListAPIKeysResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     api_keys: Optional[shared_apikeys.APIKeys] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""